"yearly" update
software page
please add an entry after the ebFRET entry for the the SMD Format software. the github page is here: https://smdata.github.io
i also have a few add ons for vbFRET and ebFRET, but perhaps we can leave these for the next time we update the website. please make a note to yourself so that you can remind me about these when we next update the site.